MILESTONE: Fixed the need to call auth_{method} manually on a timer.
It turns out that `observer`, which I had denigrated,
actually plays a key role in auth negotiation.
It is the hook that lets the user vet the host key, which is very important.
However, the way paramikojs is designed, *nothing crashed*
from me not playing along, it just silently stopped.

I found a workaround, though:
For some reason though there are *two* auth subroutines named authenticatedCallback:
 one in client.js and one in transport.js. The transport.js is more directl and manly,
 and if I just call transport.connect() a second time with authenticatedCallack=null,
 it will use the manlier callback and everything just works like a snap.